- Upgraded to OpenCV 3.4.9 - Switched to custom native libs loader - Fixed issue with launching OpenJFX 13 - Added Kotlin support
