+++ build.gradle
@@ -5,6 +5,7 @@ import org.ajoberstar.grgit.Grgit
 import org.ajoberstar.grgit.Credentials
 
 buildscript {
+    ext.kotlin_version = '1.3.61'
     repositories {
         jcenter()
         maven {
@@ -14,6 +15,7 @@ buildscript {
     dependencies {
         classpath group: 'org.codehaus.gpars', name: 'gpars', version: '1.1.0'
         classpath group: 'org.kohsuke', name: 'github-api', version: '1.+'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
     }
 }
 
@@ -23,6 +25,7 @@ plugins {
 }
 
 apply plugin: 'java'
+apply plugin: 'kotlin'
 apply plugin: 'eclipse'
 
 sourceCompatibility = 13
@@ -88,6 +91,8 @@ configurations {
 dependencies {
     jfxant files("$javaHome" + "/../lib/ant-javafx.jar")
 
+    compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.9'
+
     // FindBugs annotations to suppress warnings
     compile group: 'com.google.code.findbugs', name: 'annotations', version: '3.+'
 
@@ -128,7 +133,7 @@ dependencies {
     }
 
     //OpenCV
-    compile 'org.openpnp:opencv:3.4.2-1'
+    compile fileTree(dir: 'libs', include: '*.jar')
 
     // OSHI to collect HW and system state data
     compile group: 'com.github.dblock', name: 'oshi-core', version: '3.+'
@@ -145,6 +150,7 @@ dependencies {
 
     testCompile group: 'junit', name: 'junit', version: '4.+'
     testCompile group: 'org.hamcrest', name: 'hamcrest-core', version: '1.+'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
 }
 
 test {
@@ -735,4 +741,14 @@ release.doLast {
           "- Make Reddit release post\n" +
           "- Make Twitter release post\n" +
           "- Make release email")
+}
+compileKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
+}
+compileTestKotlin {
+    kotlinOptions {
+        jvmTarget = "1.8"
+    }
 }
