+++ src/main/java/com/shootoff/Main.kt
@@ -1,0 +1,9 @@
+package com.shootoff
+
+/*
+    JavaFX Application class was wrapped because of issue with launcher OpenJFX 11+
+    https://stackoverflow.com/a/52571719
+ */
+fun main(args: Array<String>) {
+    JFXApplication.main(args)
+}
