+++ src/main/java/com/shootoff/util/loaders/UnsupportedPlatformException.kt
@@ -1,0 +1,4 @@
+package com.shootoff.util.loaders
+
+class UnsupportedPlatformException(message: String?) : RuntimeException(message) {
+}
