+++ src/test/java/com/shootoff/TestJWS.java
@@ -1,20 +1,21 @@
 package com.shootoff;
 
-import static org.junit.Assert.*;
+import org.junit.Test;
 
 import java.util.Optional;
 
-import org.junit.Test;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertTrue;
 
 public class TestJWS {
 	@Test
 	public void testResourcesMetadata() {
-		Main main = new Main();
+		JFXApplication JFXApplication = new JFXApplication();
 
 		String metadataXML = "<?xml version=\"1.0\" encoding=\"UTF-8\"?>\n"
 				+ "<resources version=\"1.1\" fileSize=\"400319\" />";
 
-		Optional<Main.ResourcesInfo> ri = main.deserializeMetadataXML(metadataXML);
+		Optional<JFXApplication.ResourcesInfo> ri = JFXApplication.deserializeMetadataXML(metadataXML);
 
 		assertTrue(ri.isPresent());
