+++ src/test/java/com/shootoff/camera/TestAutoCalibration.java
@@ -1,17 +1,15 @@
 package com.shootoff.camera;
 
-import static org.junit.Assert.*;
-
-import java.awt.Dimension;
-import java.awt.image.BufferedImage;
-import java.io.File;
-import java.io.IOException;
-import java.util.Optional;
-
+import com.shootoff.camera.autocalibration.AutoCalibrationManager;
+import com.shootoff.camera.shotdetection.JavaShotDetector;
+import com.shootoff.config.Configuration;
+import com.shootoff.config.ConfigurationException;
+import com.shootoff.gui.*;
+import com.shootoff.gui.pane.ProjectorArenaPane;
+import com.shootoff.plugins.TrainingExercise;
+import com.shootoff.plugins.engine.PluginEngine;
+import com.shootoff.util.loaders.OpenCVLoader;
 import javafx.geometry.Bounds;
-
-import javax.imageio.ImageIO;
-
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -19,18 +17,14 @@ import org.junit.rules.ErrorCollector;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfPoint2f;
 
-import com.shootoff.camera.autocalibration.AutoCalibrationManager;
-import com.shootoff.camera.shotdetection.JavaShotDetector;
-import com.shootoff.config.Configuration;
-import com.shootoff.config.ConfigurationException;
-import com.shootoff.gui.CalibrationConfigurator;
-import com.shootoff.gui.CalibrationManager;
-import com.shootoff.gui.CalibrationOption;
-import com.shootoff.gui.ExerciseListener;
-import com.shootoff.gui.MockCanvasManager;
-import com.shootoff.gui.pane.ProjectorArenaPane;
-import com.shootoff.plugins.TrainingExercise;
-import com.shootoff.plugins.engine.PluginEngine;
+import javax.imageio.ImageIO;
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.File;
+import java.io.IOException;
+import java.util.Optional;
+
+import static org.junit.Assert.*;
 
 public class TestAutoCalibration implements VideoFinishedListener {
 	private AutoCalibrationManager acm;
@@ -44,7 +38,7 @@ public class TestAutoCalibration implements VideoFinishedListener {
 
 	@Before
 	public void setUp() throws ConfigurationException {
-		nu.pattern.OpenCV.loadShared();
+		OpenCVLoader.INSTANCE.loadSharedLibs();
 
 		acm = new AutoCalibrationManager(new MockCameraManager(), mockCamera, false);
